Hard-code `tboil` in `shr_wv_sat_mod`.
The boiling (actually steam) point in `shr_wv_sat_mod` is not the actual
boiling point, but an approximation used historically by the Goff &
Gratch method. Since it is part of the method definition, it's actually
more reasonable to hard-code it, rather than asking the user to supply a
boiling point that may defer from the expected value.

Testing:

     Unit tested only.

Code reviews:

     None.