Initialize `shr_wv_sat_mod` in the driver.
This commit adds namelist variables to control `shr_wv_sat_mod`, and
initializes it from the driver, allowing this code to be used in flux
calculations or by the various components.

Note that this will not change answers until some code begins to
actually use this module.

Testing:

    Test suite: Unit tests and driver (CIME) tests.
    Baseline: cesm1_4_alpha05c
    Namelist changes:
        - Added variables `wv_sat_scheme`, `wv_sat_transition_start`,
          `wv_sat_use_tables`, and `wv_sat_table_spacing` to the
          infodata namelist.

    Status: bit for bit

Code reviews:

     None.