Adjust 'divide' threshold, add notice if 'divide' found
In calculating diffusivity, the previous commit adds logic to ignore
flat and slow cells that presumed to be divides where diffusivity is
undefined.  This commit adjusts the speed threshold based on false
negatives encountered in running the dome for 500 yrs with Albany-FO.
It also adds a notice to the error log when divides are detected so if
the user encounters instabilities they are aware that a false positive
may have occurred.