Add allowableDtACFL & allowableDtDCFL output variables
In addition to these values being printed in the log file, this makes
them available as fields in an output file.  Note that the CFL
variables will only be calculated if
config_print_thickness_advection_info .or. config_adaptive_timestep
and the DCFL variable if one of those plus
config_adaptive_timestep_include_DCFL.  You can still get the fields in
the output file if you do not have those settings enabled but the values
will not be calculated.