add -nosavetiming option to create_newcase script
The defaults for SAVE_TIMING (and SAVE_TIMING_DIR), used to
control the automatic archive of performance data and
associated provenance, are defined per system in
config_machines.xml . This new option sets
SAVE_TIMING to FALSE in env_run.xml when
creating a new test, regardless of the default setting
in config_machines.xml. This is convenient when using a script
to generate multiple cases when performance data archiving
is not desirable, such as in some test suites.

It would be more general to use something like '-savetiming true'
or '-savetiming false', but SAVE_TIMING requires also the definition
of the location of the data archive (SAVE_TIMING_DIR) for anything
to happen, so the script is still dependent on the definitions in
config_machines.xml. We could add yet another parameter to set
SAVE_TIMING_DIR, but this would only be required if SAVE_TIMING
is being set to TRUE. As the only use case at the moment is to
disable SAVE_TIMING when generating multiple cases
in scripts, this generality is not yet, and may never be,
important. For individual cases, it is simpler to just edit
env_run.xml, either manually or with ./xmlchange.