Add new ACME version of create_test
Currently, just a wrapper around the CESM version of this
script that presents an improved interface.

It also allows for convenient creation of sets of tests to run
on the fly.

Uses posix compliant options.

Change other acme tools to use this create_test instead of old one.

Eventually, I'd like this script to fully replace CESM's create_test.

[BFB]