Changes to get acme scripts regression tests working again
Recent changes broke the tests are were not caught due to a bug
that caused the regression test report to get sent to the ACME_test
project instead of ACME_Climate.

Add --force option to bless_test_results so it will work when
being scripted (non-interactive).

Regression tests now preserve any DartConfiguration files instead
of letting wait_for_tests blow it away. This was the cause of results
being sent to the wrong dashboard.

[BFB]