Deactivate curvature code with cpp directives
Code for handling curvature terms was previously active, though controlled
with a Fortran parameter. This code is now activated by defining the
cpp macro CURVATURE (which can be done, e.g., in the definition of
MODEL_FORMULATION in the top-level Makefile), though with this commit,
the curvature code is now inactive.