Commit of nearly-functional passing of ice sheet mask to sea ice model.
This commit provides near functionality of the ability to pass the ice
sheet extent mask (and, potentially, any ice sheet field) to the sea ice
model.  This required generation of new scripts and code to allow for Sg2i
and Fg2i coupling functionality (previously, only Rg2i existed, to allow
for potential ice flux calving to an explicit iceberg model.

Some more work will be required to separate the coupling flags, so that,
e.g, ice-shelf-specific non-calving fields (specifically ice sheet
extent mask, here) are not dependent upon the existing iceberg coupling
flags, which were meant to determine whether icebergs are routed to the
ocean or the sea ice.  This work will predominantly go in
ccsm_comp_mod.F90, and should be the topic of the next commit after
this.

Changes in this commit:
-*namelist* & *.xml files: infrastructure for Sg2i/Fg2i mapping file
read-in.
-seq_flds_mod.F90: added seq_flds_add call so that
Sg_icemask_coupled_fluxes field is passed to ocean and sea ice (in
addition to the land)
-ccsm_comp_mod.F90: some temporary commenting of if loops, to bypass
coupling flag infrastructure and test glc->ice field passing.
-prep_ice_mod.F90: initialization of mapper_Sg2i.  Note, an equivalent
Fg2i mapper not initialized (despite being allowable in the setup
scripts).  Can set this up later if needed.
-prep_ocn_mod.F90: removed indexes to g2x/x2o_Sg_topg (not needed, was
only used for a testing print statement)