Update mpas_dmpar to be private by default
This commit migrates the mpas_dmpar module to be private by default
instead of public. In addition it defines the public interface for the
mpas_dmpar module.

Two places made use of the mpas_sort module's subroutines which are no
longer publicly available through mpas_dmpar after this change. Fixes to
those are included as well.