Fixing duplicate_field to check isActive
Previously, the duplicate_field routines would always attempt to copy
the array / scalar part of a field from the source to the destination
field, even if one of the two fields was not active
(i.e. `field % isActive = .false.`)

This commit updates the duplicate_field routines to check the isActive
attribute before copying data.