Add `override` to incremental definitions
Some build systems ignore the += operator on previously defined
variables if the `override` option is missing. This commit adds
`override` to variables that need to be incrementally defined.

The specific issue this commit fixes is that CPPDEF's is not added into
CPPFLAGS on some build systems, which causes the -DUNDERSCORE definition
to not be added to CPPFLAGS, and all of the c code within MPAS is not
built properly.