Add `shr_RandNum_mod`.
`shr_RandNum_mod` is the Fortran interface for a collection of random
number generator codes, collected and optimized by Chris Kerr.

This commit simply adds the files and a benchmark/test, with some
modifications to the code and directory structure to simplify the build
and for portability. The main deviations from the original are:

 1) The input to create a random number generator is now a single 2D
    seed array, rather than a set of optional arguments for different
    RNG methods.

 2) The Makefile is modified to work with the changed build directory
    structure and on different compilers.

As currently implemented, the new code will be linked in to the
executable, but there are no users of the code until RRTMG is modified to
leverage `shr_RandNum_mod`.

Testing:

    Test suite: System tests on Yellowstone/Intel,
                benchmark test run on Intel, PGI, GNU, NAG
    Baseline: None used.
    Namelist changes: None.
    Status: Presumably bit-for-bit (linking in an unused library shouldn't
            change answers).

Code reviews:

     None (except dialog between Chris and Sean Santos).