Increasing ShortStrKIND length
ShortStrKIND is only used for defining the names of alarms on a given
clock. Every stream generates two alarms by default that are the name of
the stream with the prefixes _input and _output. If a stream name is
longer than 25 characters, the alarm infrastructure doesn't work
correctly and an error message that is not useful is thrown.

This commit increases the length of alarm names, and also adds more
useful error messages when an alarm name is provided that is longer than
the acceptable length.