XML file changes to topo grid name for CONUS case
It was discovered that the topo file used for the CONUS test case was
erroneous and had the too coarse of topogrpahy in the CONUS region.
A new CONUS topo grid has been added to the ACME inputdata repo, and
when building this case, the user will download the new topo file.
 On branch eroesler/ATM/RRM-conustopofilefix
 Changes to be committed:
	modified:   ../../models/atm/cam/bld/namelist_files/namelist_defaults_cam.xml

[NML]

AG-296