Fixing an issue that prevented bit reproducibility
Previously, surface tracer fluxes when performing a run with restoring
would only be computed over nCellsSolve. This meant cells outside of
nCellsSolve had a restoring flux of 0 instead of the correct flux for
their tracer values.

This commit changes restoring to compute fluxes over nCells to ensure
correct values are provided for all cells in a block.