Fixing timer names in timer_stop calls
In the analysis members and the TEMPLATE, the timer names in the
timer_stop calls were different than the timer names in the timer_start
calls. This can cause issues when trying to profile the code, in
addition to providing incorrect timer information.

This commit fixes the analysis members so the names in both calls are
the same.