Adding additional options for timer libraries.
This commit adds additional options for timer libraries within the
mpas_timer module.

Previously, there were native timers, and hooks to call TAU timers. This
commit adds support for gptl in a stand-alone capability (with the GPTL
environment variable pointing to where gptl is installed), and perf_mod
times that are used in coupled models like ACME and CESM.

When building the model, a new option "TIMER_LIB" can be used to control
which timer interface you want. It can take the following values:
 - TIMER_LIB=native - (Default) uses the old style native MPAS timers
 - TIMER_LIB=tau - Disables the native timers, and only uses TAU for timers
 - TIMER_LIB=gptl - Disables the native timers, and only uses GPTL for timers

There isn't a way in stand-alone MPAS to use the perf_mod timers, but
when building in a coupled model -DMPAS_PERF_MOD_TIMERS can be added to
the CPPFLAGS to use these timers.