Add new performance module and simplified timing code
This commit adds a new 'mpas_performance' module that can be used to read and
report event counters using PAPI (compile with USE_PAPI=true and the PAPI env
variable set to the installation path of the PAPI libraries).

Before making any calls to the performance module, the routine mpas_performance_init()
must be called, and the mpas_performance_finalize() routine should be called after
all other calls to the performance module. Counting cycles and FP ops for a section
of code is accomplished by adding calls like the following around the code in question:

    call mpas_performance_start()

    !
    ! code ...
    !

    call mpas_performance_stop('code section name')

Also included in this commit is a cleanup of the C timing code in timer.c, for
which the mpas_performance module includes an explicit interface. To use the
timing code based on gettimeofday(), one must include the mpas_performance module
to get the interface:

    use mpas_performance

and declare variables to hold the integer seconds and microseconds returned by
the stop_timer routine:

    use iso_c_binding, only : c_int
    integer (c_int) :: secs, u_secs

and add timing calls around the code in question:

    call start_timer(0)

    !
    ! code ...
    !

    call stop_timer(0, secs, u_secs)
    write(0,'(a,i2,a1,i6.6)') 'Elapsed time: ', secs, '.', u_secs

There are ten available timers, 0 .. 9, that can be used as the first argument to
the start_timer and stop_timer routines. The overhead for a pair of timer calls
(start_timer followed by stop_timer) is around one microsecond (on a 2.6 GHz Xeon E5).