Convert threaded subroutine calls in dynamics to use mpas_threading module
Prior to the availability of the mpas_threading module, the threaded
subroutine calls in the MPAS-Atmosphere dynamics created an OpenMP parallel
region around each subroutine call with private variables for the element
start and end loop bounds, obtained the number of OpenMP threads and the
ID for each thread, and computed these bounds before calling the subroutine.

Using the mpas_threading module, all such parallel subroutine calls in
the dynamics have been updated so that a simple OpenMP "PARALLEL DO"
is needed, with the do-loop over the number of threads in the block, e.g.,

!$OMP PARALLEL DO
   do thread=1,block % nThreads
      !!! subroutine call !!!

The element loop bounds in the subroutine argument lists are simply taken
from those that were pre-computed in each block by the mpas_threading
module.