wait_for_tests: By default, do not fail a test if only a memcomp check failed
Adds an option to fail memcomp failures.

On blues, the memcomp checks are not reliably passing, even when generating
results and comparing against the same commit. This was causing a lot of
problems in the nightlies.

Going forward, only tests that are specifically for performance
should check throughput and memory.

[BFB]