Allocate a "garbage element" for temp arrays used in mpas_atm_time_integration.F
For a number of subroutines in mpas_atm_time_integration.F, we allocate block-sized
arrays for cell, edge, or vertex elements outside the threaded calls to these
subroutines. In order to ensure that loops in these routines don't access memory
outside the allocated bounds of these arrays, while still avoiding if-tests in
the loops, we need to allocate "garbage elements" as is done with field arrays
defined in the Registry.xml file; i.e., we allocate a cell-based array with
(nCells+1) elements, an edge-based array with (nEdges+1) elements, etc.