Move module variables used as loop variables to stack variables in subroutines
Many of the modules in src/core_atmosphere/physics/ declared module variables
(i, j, k, etc.) as module variables, and these module variables were then used
as loop variables in subroutines of those modules. However, when the calls to
such subroutines are threaded, race conditions on the updates of these loop
variables arise.

In order to thread calls to subroutines in the MPAS physics, this commit moves
declarations of module variables that were used as loop variables into
the individual subroutines that require them.