Declare 'nlayers' as a local variable in the rrtmg_lwrad() subroutine.
This variable is also declared as a module variable, which causes trouble
for threaded calls to rrtmg_lwrad(). This worked fine in WRF, with its
pressure coordinate, since nlayers had the same value for all columns;
however, it doesn't work in MPAS when each column needs to recompute the
number of additional layers to add to a column.