Edison env cleanup
Do a purge in the env mach file to remove differences between different
user's modules. We need to ensure all ACME users get a consistent env on
edison for running ACME.

For ACME to work on edison, the module cray-libsci must NOT be loaded;
however, the intel PrgEnv module loads the cray-libsci module, so
cray-libsci must be unloaded after the PrgEnv load. The problem is,
the following sequence is broken on edison:

    module load PrgEnv-intel
    module unload cray-libsci
    module purge
    module load PrgEnv-intel
    module unload cray-libsci

The second time through, the cray-libsci is not successfully unloaded
and remains in the env to cause trouble. The old system that did not
use a purge could sometimes worked dependeding on the users' module
setup.

To address this:

    * Env_specific.edison now has a guard, like a C++ include file, so that it is not sourced multiple times
    * Env_specific.edison does a purge at the beginning of the file so that all users experience get the same env when using ACME