Allow the MPAS drivers to overwrite the log units
Previously, each MPAS component shared a log unit number (stderrUnit and
stdoutUnit). If multiple MPAS components are running on the same PE, the
problem is that all of the output ends up getting redirected to an
inconsistent log file.

This commit changes it so the ACME drivers for MPAS components will
ensure their log units point to the correct place at each phase of the
run.