Remove ERS_Ly21.f09_g16.TG test
The ERS_Ly21.f09_g16.TG test tests CISM with data LND.  The changes in this
branch generalize the GLC component so that the existing usage of CISM
no longer works.  However, the ACME Land Ice team has agreed that
removal of this test is acceptable since the corresponding
generalization of the CISM implementation will be added in a
separate branch, and this test can be restored at that point if desired.

[Non-BFB]