Add the ability to pull input and output intervals from other streams
This commit adds the ability for a stream's input and output intervals
to refer to the input and output intervals of other streams. In this
case, the syntax for doing so looks like:

input_interval="stream:stream_name:interval_name"

An example would be:

input_interval="stream:restart:input_interval"

Note, input intervals are not required to refer to input intervals (i.e.
they can refer to output intervals). The same is true for output
intervals.