Allow vector reconstruciton to provide reproducibile results
Previously, the vector reconstruction routines would only perform
reconstruction over owned cells within a block. This commit allows an
optional argument that can be used to reconstruct over all elements
within a block.

If the old version of vector reconstruction was used for anything where
the reconstructed fields were used, they needed a a halo exchange to be
correct over the full block. This commit also allows the removal of
these halo exchanges.