Allow deactivated fields to have missing dimensions
If a field is deactivated by a package, and one of their dimensions is
not defined, previously framework would throw an error. However, the
dimension is not required to be defined, since the the field isn't used.
This commit changes the dimension validation logic to allow deactivated
fields to have missing dimension values.