Set GLC grid to be the same as LND grid for f09_g16
Previously ACME would internally set the GLC grid to be equal to the LND
grid.  The grid specified as the GLC grid was actually what was called
the ICESHEET_GRID and not used internally by ACME.  This branch removes
that restriction, but that change caused the X test to fail because now
the GLC grid was null.  This change to the f09_g16 compset sets the GLC
grid as the LND/ATM grid (0.9x1.25) so the X test will continue to work.

Note that 0.9x1.25 is not a meaningful GLC grid and this is just a way
to continue allowing the X compset/test to function.  This GLC grid
should eventually be replaced with a proper GLC grid (such as
mpas.gis20km) but for that to work, mapping files are needed between
that grid and the LND grid in this compset.