Fixing single threaded builds on Cetus
This commit fixes build time issues with single threaded builds
on Cetus. Without the fix a single threaded build would fail
due to mixing of single threaded and multi threaded (openmp)
compiler options.
The error does not show up in acme developer or integration builds
because typically on Cetus all builds are multithreaded. The
user will notice the build error if BUILD_THREADED is set to false
(in env_build.xml) and the number of threads for every component
is set to 1 (in env_mach_pes.xml).

[BFB]
Issue #288