Reject use of OpenMP in NAG builds
Since using OpenMP+NAG is not possible, and people have wasted a lot of
time debugging confusing error messages as a result, it's best to reject
such cases. This is done at cesm_setup time, since that's the earliest
point where we know with confidence whether the build will be threaded.

Testing:

    Test suite: NAG prealpha on hobart
    Baseline: cesm1_4_alpha07b
    Namelist changes: None
    Status: bit for bit