Ensure garbage elements are giving a global index of 0
Previously the garbage elements (cells, edges, and vertices) were not
initialized to zero. This caused some runs to have the garbage elements
be filled with large integer values, and visualization tools would be
unable to read the connectivity information.

When remapping the connectivity fields (prior to writing out mesh
information) the garbage elements (i.e. anything that had a missing
element) could be filled with a large value breaking the output mesh.

The issue this commit fixes only shows up with meshes that have
boundaries.