changed vertical interpolation ordering of arrays
Changed the vertical interpolation scheme in the eliassen_palm AM so
that now it expects monotonically increasing reference array.
The reason for this change is that most calls to this routine were
sending -potentialDensityMidRef, so the compiler had to make a scratch
variable. In order to clean this up, I decided to change the ordering of
the arrays expected as arguments, instead of having to assign temporary
scratch variables everywhere to be sent to the interpolation routine.

The vertical interpolation routine was originally developed with the
idea of sending it a reference buoyancy variable, not density.
In hindsight, the vertical interpolation should expect a reference
density variable to interpolate to. For nonlinear equations of state,
this reference density variable can be neutral density.