Separating out the redi portion of hmix for gm
This commit separates out the redi portion of hmix that is used when GM
is turned on. Previously this was incorporated within the del2 module,
and the logic to control if redi was used or not was confusing.

In addition, this commit updates the redi hmix module to be bit
reproducible, which it was not previously.