Adding an isActive check for halo exchanges
This commit adds a test within halo exchanges to see if a field is
active before performing a halo exchange. This allows package variables,
which may or may not be active to be passed into the halo exchange
routine without testing if the field is active first.

When built in debug mode, the halo exchanges write a message stating the
the halo exchange was skipped.