Fixed a minor bug in diagnostic output
The code was writing output to the error log that the requested diagnostic cell is not on the processor
(even when it was on the processor) whenever verboseInit = .false.

Now the output correctly states whether the requested cell is on the processor, provided that
verboseInit = .true.  For verboseInit = .false., there is no message.