registry for MLD was modified to include latCell and lonCell in stream for easy plotting
the analysis member had a few logical errors, such as recomputing the MLD even after it was found at the previous depth level.  It was also modified to include constraints that the MLD fall between the two values of zMid of interest.  This prevents extrapolation producing bad values.  This was evident in very coarse simulations where the temperature changed quickly between the first and second levels