warn if compute on startup and restart both true
If compute on startup and a restart are both requested, there could be
problems if the ensemble averaged loaded in the AM restart were already
calculated for the state being loaded in the forward model restart file.
This problem could occur if the output interval of the AM restart file
is different to the compute interval.