Change way scalar fields are extended at floating boundaries in Interface
Change implementation of fields (elevation, thickness, basal friction, etc.)
extensions at floating boundaries.   We now extend the fields only on floating
boundary cells that do not have ice, looking at adjacent cells that have ice.
Among adjacent cells with ice we pick the one with smallest elevation and copy
the fields from that cell to the boundary one.  In order to do that we need to
pass the cellMask to the velocity solver interface.  We do not check whether the
cell mask changes in time or not, assuming for now that it changes only when the
vertexMask changes.