IR fixes to handle 2D fields and prevent negative mass
This commit includes two major fixes:

(1) Modified treatment of mass/tracer fields that are physically 2D but formally 3D

In principle, MPAS-CICE allows both 2D and 3D mass and tracer fields, but currently
all tracer fields are 3D.  Fields such as ice concentrations and thickness, although physically 2D,
have a first (level) dimension of 1.  I call these "faux-3D" fields.

Until now, the IR code has assumed that if a child tracer (e.g., enthalpy) is 3D and its parent tracer is
also 3D, then the parent has the same number of levels as the child.  This assumption is incorrect
for tracers like enthalpy that are physically 3D and had faux-3D parents.

One solution would have been to add logic in the main IR subroutines to handle the faux-3D fields.
However, the new logic would have been fairly complex and would be obsolete as soon as the faux-3D
fields become 2D (as is planned for later).

An easier solution, which I implemented, is to check for faux-3D fields in the tracer setup subroutine,
set_tracer_array_pointers.  This subroutine assigns 2D pointers to elements (1,:,:) of the faux-3D fields.
Thereafter, the IR routines treat these fields as 2D.
At some point, when the faux-3D fields become formally 2D, the extra logic can easily be removed.

In this way I was able to exercise and verify much of the logic for 2D fields for the first time.
I found and fixed bugs in subroutine limit_tracer_gradient_2d and check_tracer_monotonicity, and in some diagnostic output.
I also added some logic in limit_tracer_gradient_3d to distinguish between 2D and 3D parents.

(2) Modified x/yVertexOnEdge coordinates to prevent negative masses

The IR geometry calculations involve computing triangle vertices in edge-based coordinates.
I found that in some cases on a spherical mesh, a triangle that is fluxed across two edges
(first edge 1 and then edge 2) could be computed to have a larger area when fluxed across
edge 2 than when fluxed across edge 1, because angles between cell edges and departure vectors
can differ depending on which edge the triangle is crossing.  This can lead to more mass leaving a cell
across edge 2 than entering across edge 1, resulting in negative mass.

The solution was to change the way I computed x/yVertexOnEdge, the coordinates of the vertices near
a given edge.  Of these neighbor vertices (6 for hexes and 8 for quads), two vertices lie on the edge
itself (call these the nearest vertices), and the remaining vertices are more distant, lying at the
opposite end of the edges that share a single vertex with the main edge.
I made sure that the distance from an edge midpoint to a more distant vertex is exactly equal to a
sum of distances between edge midpoints and nearest vertices.  The distance to nearest vertices is
computed once and for all for each edge, and then neighboring edges refer to this same distance.

I also modified some edge-to-cell transformations to make them area-preserving.
Previously, there was a slight change in area when a triangle was transformed from edge-based
coordinates to neighboring cell-based coordinates.  Now, a triangle computed in an edge-based
coordinate system is simply translated in space by the distance between the edge and the
neighboring cell center. This distance is given by the value of new coordinates xCellOnEdge and yCellOnEdge
that are computed at initialization for each neighboring cell of each edge.
Note: I made the changes involving x/yCellOnEdge before figuring out the x/yVertexOnEdge solution above.
They didn't fix the negative areas, but I think they are sensible changes.

For a spherical-hex test problem, I verified that with the new transformations, answers are very similar
to those with the old transformations (but without the rare negative areas); ice areas after a few time steps
typically differ by 10^(-4) or less.

For a planar-quad test problem, answers are BFB with the new transformations.
For a planar-hex problem, answers are the same within machine roundoff.

I also verified for the planar-quad, planar-hex and spherical-hex problems that results
on 4 cores (on a Mac) are the same (BFB) as on 1 core, as was the case before this round of changes.

I did a bit of minor code cleanup along the way.