Check verticesMask when interpolating velocity to edges in Interface
When computing normal velocity, we need to make sure that the vertex (triangle in
the dual Delaunay mesh) we use for the interpolation belongs to the verticesMask.

For interior edges this is never an issue, but in some situations we may
be interpolating velocity to edges on the boundary of the FEM mesh, in
which case we need to make sure we are using a valid FEM triangle!