Check for valid triangles when setting vertexMask
With this change vertices are only considered 'valid' if they have three valid neighboring
cells (i.e., the cells exist in the mesh).  This allows external dycores to use the
vertexMask to get information about triangles in the Delaunay triangulation.

This is done in a way which does not assume vertexMask==3, but the triangular dual mesh
is the intended purpose for this.