Moved several land-ice related fields to diagnostics
The boundary-layer T and S, the heat and salt transfer
velocitities, the top drag and the friction velocity
are now computed as part of diagnostics.  This ensures
that they are at the right time level when they are
used to compute fluxes and better separates fields that
will always be computed in MPAS-O from those that will
sometiems be computed in the coupler.