Added support for a spatially variable attenuation coeff
Under land ice, this coefficient is likely to have a larger value
(partly because of the compressed layers) than in the open ocean.

The attenuation coefficient is computed as a diagnostics at every
time step, taking the default as before if land-ice fluxes are
turned off or wherever landIceFraction == 0.  The attenuation
coefficient transitions linearly with the landIceFraction to
its value under land ice.