Add compset for Albany solver for MPASLI
This creates separate GLC component names: MPASLISIA, MPASLIALBFO to be
used in compset descriptions depending on which version of MPASLI is
desired.

These compset definitions control a MPASLI_USE_ALBANY variable set in
config_compsets.xml.

In a separate commit that variable will be used to control the MPASLI
build within the mpasli component.