Minor IR mods for test cases
I made some minor changes in the IR code:

(1) In mpas_cice_initialize.F, I modified the criteria for the initial circle of ice
in the spherical test cases.  Previously, the initial ice-covered region was more elliptical
than circular.

(2) I modified the zapSmallAreas subroutine to zero out the tracer arrayMask in addition
to the array itself.  If the mask is not zeroed out, then the code will erroneously check
for tracer monotonicity and give an error message.

(3) I increased	the threshold value for	a fatal	tracer conservation error from 10^(-11) to 10^(-8).
When the code is tested later with more realistic tracers, I will experiment with different values.