Change references to config_density0 to rho_sw
This commit updates all references to config_density0 to use rho_sw
instead. rho_sw is a constant that the coupler defines when run in a
coupled model, and this allows all use of a reference density to use
consistent values of the reference density.

Additionally, this commit moves config_density0 from the
pressure_gradient namelist record to a new record named ocean_constants.

Finally, this commit performs some whitespace clean up and adds _RKIND
to some constants.