Added cam tests for polar mods, MAM4 and prescribed aerosols
This commit adds tests for testing polar mods, MAM4 and prescribed
aerosols in the acme testing suites. Polar mods and MAM4 are added to
acme developer test suite and a test for prescribed aerosols is added
to the acme integration test suite (please note that acme developer is
a subset of acme integration test suite). I have added a compset for
the COSP simulator but the COSP test is not functional at this moment.

[BFB]
AG-350