IR fixes for spherical quad tests with realistic forcing
For the first time, I ran the IR code with realistic forcing and with the
velocity solver and column physics turned on.  Results look good.
I fixed several issues that came up along the way:

(1) Added a check for negative areas

I found an error in the mpasMeshConverter tool that led to negative areas on the
spherical-quad meshes (with land) for ocean cells that have a land cell at one corner.
I will work with Doug to fix the bug. In the meantime, Mark Petersen provided
revised spherical-quad grid files with the correct areas.

In the IR code I added a check for non-positive values of areaCell. The code now aborts
with an error message if any cell has a negative or zero area.

(2) Turned off transport for edges without 2 cell neighbors

I set remapEdge = 0 for edges that have fewer than 2 cell neighbors.
Previously, the code was trying to compute fluxes across edges with only 1 cell neighbor
(and the other neighbor non-existent).

(3) Added a roundoff threshold for tracer monotonicity

Previously, the tracer monotonicity check did not allow for roundoff error.
Now it allows errors of up to 10^(-11) relative to the tracer magnitude.

In addition, I set extendedMaxMin = .true., in order to do a halo update of the
local max and min before checking for monotonicity.

Both the roundoff threshold and the halo update are also done in standard CICE
and are needed to avoid small, rare monotonicity violations.

(4) Added volume/thickness conversions

I added subroutines to convert from ice and snow volume to thickness at the start
of the IR computation, and from thickness to volume at the end. This was not needed
previously, because iceVolume and snowVolume were dummy tracers with no effect elsewhere.
But with the column physics turned on, it is necessary to convert area*tracer products
to true tracers for purposes of IR.

In order to call these subroutines, I added to the Registry an IR work array, workCategoryCell,
of dimension (nCategories, nCells).

With these changes, I ran MPAS CICE to completion on the gx1 grid (for 1 model day)
and on the gx3 grid (for 10 model days) using the LY forcing provided by Adrian,
using IR with the velocity solver and column physics turned on.  I looked at some
output fields and didn't see anything amiss.