analysis member output at startup w/o computation
This allows analysis members, like LIGHT, to output
analysis member input data at startup and avoid a
computational step (e.g., output the initial particle
positions within LIGHT).  This commit adds generality and a warning
is still issue if output at start up and computation at
startup are not both selected.  This follows the
premise that, in general, it should be the user responsibility to make
sure that computations are enable at startup because otherwise
analysis member functionality is adversely limited for the general
case as discussed above using LIGHT as a use case.