Fixed IR gradients on rotated spherical grids
The IR code constructs gradients at cell centers based on the normal gradients at edges.
The resulting gradient vector lies in 3D Cartesian x/y/z space.
It is then rotated to the local tangent plane of the cell.

When IR is run on a rotated mesh (where each Pole is rotated 90 degrees to the equator), the
gradient vector must be rotated by 90 degrees before being transformed to the local tangent plane.
With this commit I am adding code to do the rotation.

This commit is answer-changing for runs on a sphere with a rotated mesh.